#10 Select and Insert RegEx implemented in Parse function

+++ DBManager/Parser/MiniSQLParser.cs
@@ -11,9 +11,11 @@ namespace DbManager
     {
       //TODO DEADLINE 2
       //SELECT ColumnName[,ColumnName…] FROM Table [WHERE Condition]
-      const string selectPattern = null;
+      const string selectPattern = @"^SELECT\s+(?<columns>\w+)\s+FROM\s+(?<table>\w+)\s+WHERE\s+(?<column>\w+)\s?(?<operator>[<=>])\s?(?<value>.+)$";
+      var select = new Regex(selectPattern, RegexOptions.IgnoreCase);
       //INSERT INTO TableName VALUES (LiteralValue[,LiteralValue, …])
-      const string insertPattern = null;
+      const string insertPattern = @"^INSERT\s+INTO\s+(?<table>\w+)\s+VALUES\s+(?<values>\w+)$"; ;
+      var insert = new Regex(insertPattern, RegexOptions.IgnoreCase);
       //DROP TABLE TableName
       const string dropTablePattern = @"^DROP\s+TABLE\s+(?<table>\w+)$";
       var dropTable = new Regex(dropTablePattern, RegexOptions.IgnoreCase);
@@ -22,8 +24,10 @@ namespace DbManager
       //And then, an execution error should be given if a CreateTable without columns is executed
       //CREATE TABLE TableName (ColumnName DataType[,ColumnName DataType... ])
       const string createTablePattern = null;
+      var createTable = new Regex(createTablePattern, RegexOptions.IgnoreCase);
       //UPDATE TableName SET ColumnName=LiteralValue[,ColumnName=LiteralValue,…] WHERE Condition  
       const string updateTablePattern = null;
+      var updateTable = new Regex(updateTablePattern, RegexOptions.IgnoreCase);
       //DELETE FROM TableName WHERE Condition
       const string deletePattern = @"^DELETE\s+FROM\s+(?<table>\w+)\s+WHERE\s+(?<column>\w+)\s?(?<operator>[<=>])\s?(?<value>.+)$";
       var deleteFrom = new Regex(deletePattern, RegexOptions.IgnoreCase);
@@ -50,28 +54,73 @@ namespace DbManager
       //If there is no match, it means there is a syntax error. We will return null.
       var trimmedQuery = miniSQLQuery.Trim();
 
-      var dropTableMatch = dropTable.Match(trimmedQuery);
-      var deleteFromMatch = deleteFrom.Match(trimmedQuery);
+      var selectMatch = select.Match(trimmedQuery);
+      if (selectMatch.Success)
+      {
+        var table = selectMatch.Groups["table"].Value;
 
+        var unsplittedColumns = selectMatch.Groups["columns"].Value;
+        var splittedColumns = unsplittedColumns.Split(',');
+        var columns = new List<string>();
+        foreach (string column in splittedColumns) columns.Add(column);
+
+        var columnName = selectMatch.Groups["column"].Value;
+        var op = selectMatch.Groups["operator"].Value;
+        var value = selectMatch.Groups["value"].Value;
+        var condition = new Condition(columnName, op, value);
+
+        return new Select(table, columns, condition);
+      }
+
+      var insertMatch = insert.Match(trimmedQuery);
+      if (insertMatch.Success)
+      {
+        var table = insertMatch.Groups["table"].Value;
+
+        var unsplittedValues = insertMatch.Groups["columns"].Value;
+        var splittedValues = unsplittedValues.Split(',');
+        var values = new List<string>();
+        foreach (string value in splittedValues) values.Add(value);
+
+        return new Insert(table, values);
+      }
+
+      var dropTableMatch = dropTable.Match(trimmedQuery);
       if (dropTableMatch.Success)
       {
         var table = dropTableMatch.Groups["table"].Value;
         return new DropTable(table);
       }
+
+      var createTableMatch = createTable.Match(trimmedQuery);
+      if (createTableMatch.Success)
+      {
+        return null;
+      }
+
+      var updateTableMatch = updateTable.Match(trimmedQuery);
+      if (updateTableMatch.Success)
+      {
+        return null;
+      }
+
+      var deleteFromMatch = deleteFrom.Match(trimmedQuery);
       if (deleteFromMatch.Success)
       {
         var table = deleteFromMatch.Groups["table"].Value;
+
         var column = deleteFromMatch.Groups["column"].Value;
         var op = deleteFromMatch.Groups["operator"].Value;
         var value = deleteFromMatch.Groups["value"].Value;
         var condition = new Condition(column, op, value);
+
         return new Delete(table, condition);
       }
 
-        //TODO DEADLINE 4
-        //Do the same for the security queries (CREATE SECURITY PROFILE, ...)
+      //TODO DEADLINE 4
+      //Do the same for the security queries (CREATE SECURITY PROFILE, ...)
 
-        return null;
+      return null;
            
     }
